A bunch of rather random changes
