+++ src/main/java/me/william278/huskhomes2/api/events/PlayerChangeWarpDescriptionEvent.java
@@ -4,9 +4,10 @@ import me.william278.huskhomes2.teleport.Warp;
 import org.bukkit.entity.Player;
 
 /**
- An event, fired when a player changes the description of a warp.
+ * An event, fired when a player changes the description of a warp.
  */
 public class PlayerChangeWarpDescriptionEvent extends PlayerWarpUpdateEvent {
+
     private final String oldDescription;
     private final String newDescription;
