+++ src/main/java/me/william278/huskhomes2/api/events/PlayerRelocateHomeEvent.java
@@ -9,8 +9,8 @@ import org.bukkit.entity.Player;
  */
 public class PlayerRelocateHomeEvent extends PlayerHomeUpdateEvent {
 
-    final TeleportationPoint oldTeleportationPoint;
-    final TeleportationPoint newTeleportationPoint;
+    private final TeleportationPoint oldTeleportationPoint;
+    private final TeleportationPoint newTeleportationPoint;
 
     /**
      * An event, fired when a player relocates a home
