+++ src/main/java/me/william278/huskhomes2/commands/SethomeCommand.java
@@ -1,8 +1,8 @@
 package me.william278.huskhomes2.commands;
 
-import me.william278.huskhomes2.data.DataManager;
 import me.william278.huskhomes2.MessageManager;
-import me.william278.huskhomes2.SettingHandler;
+import me.william278.huskhomes2.data.DataManager;
+import me.william278.huskhomes2.teleport.SettingHandler;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
