+++ src/main/java/me/william278/huskhomes2/teleport/TeleportRequestHandler.java
@@ -1,9 +1,12 @@
 package me.william278.huskhomes2.teleport;
 
 import me.william278.huskhomes2.HuskHomes;
-import me.william278.huskhomes2.integrations.VanishChecker;
 import me.william278.huskhomes2.MessageManager;
 import me.william278.huskhomes2.PluginMessageHandler;
+import me.william278.huskhomes2.data.DataManager;
+import me.william278.huskhomes2.integrations.VanishChecker;
+import me.william278.huskhomes2.integrations.VaultIntegration;
+import net.md_5.bungee.api.ChatColor;
 import net.md_5.bungee.api.chat.ClickEvent;
 import net.md_5.bungee.api.chat.ComponentBuilder;
 import net.md_5.bungee.api.chat.HoverEvent;
@@ -11,13 +14,17 @@ import net.md_5.bungee.api.chat.TextComponent;
 import net.md_5.bungee.api.chat.hover.content.Text;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
+import org.bukkit.plugin.Plugin;
 
 import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Map;
+import java.util.Set;
 
 public class TeleportRequestHandler {
 
     // Target player and teleport request to them hashmap
-    public static HashMap<Player, TeleportRequest> teleportRequests = new HashMap<>();
+    public static Map<Player, TeleportRequest> teleportRequests = new HashMap<>();
 
     private static void sendTeleportRequestCrossServer(Player requester, String targetPlayerName, String teleportRequestType) {
         String pluginMessage = teleportRequestType + "_request";
@@ -29,7 +36,7 @@ public class TeleportRequestHandler {
         PluginMessageHandler.sendPluginMessage(replier, requesterName, pluginMessage, replier.getName() + ":" + accepted);
     }
 
-    public static TextComponent createButton(String buttonText, net.md_5.bungee.api.ChatColor color, ClickEvent.Action actionType, String command, String hoverMessage, net.md_5.bungee.api.ChatColor hoverMessageColor, Boolean hoverMessageItalic) {
+    private static TextComponent createButton(String buttonText, ChatColor color, ClickEvent.Action actionType, String command, String hoverMessage, ChatColor hoverMessageColor, Boolean hoverMessageItalic) {
         TextComponent button = new TextComponent(buttonText);
         button.setColor(color);
 
@@ -42,17 +49,17 @@ public class TeleportRequestHandler {
         // Send the "Accept" or "Decline" response buttons to the player who has received a request
         // Options text
         TextComponent options = new TextComponent(MessageManager.getRawMessage("tpa_request_buttons_prompt"));
-        options.setColor(net.md_5.bungee.api.ChatColor.GRAY);
+        options.setColor(ChatColor.GRAY);
 
         TextComponent separator = new TextComponent(MessageManager.getRawMessage("list_item_divider"));
-        separator.setColor(net.md_5.bungee.api.ChatColor.GRAY);
+        separator.setColor(ChatColor.GRAY);
 
         // Build the components together
         ComponentBuilder teleportResponses = new ComponentBuilder();
         teleportResponses.append(options);
-        teleportResponses.append(createButton(MessageManager.getRawMessage("tpa_accept_button"), net.md_5.bungee.api.ChatColor.GREEN, ClickEvent.Action.RUN_COMMAND, "/tpaccept", MessageManager.getRawMessage("tpa_accept_button_tooltip"), net.md_5.bungee.api.ChatColor.GRAY, false));
+        teleportResponses.append(createButton(MessageManager.getRawMessage("tpa_accept_button"), ChatColor.GREEN, ClickEvent.Action.RUN_COMMAND, "/tpaccept", MessageManager.getRawMessage("tpa_accept_button_tooltip"), ChatColor.GRAY, false));
         teleportResponses.append(separator);
-        teleportResponses.append(createButton(MessageManager.getRawMessage("tpa_decline_button"), net.md_5.bungee.api.ChatColor.RED, ClickEvent.Action.RUN_COMMAND, "/tpdeny", MessageManager.getRawMessage("tpa_decline_button_tooltip"), net.md_5.bungee.api.ChatColor.GRAY, false));
+        teleportResponses.append(createButton(MessageManager.getRawMessage("tpa_decline_button"), ChatColor.RED, ClickEvent.Action.RUN_COMMAND, "/tpdeny", MessageManager.getRawMessage("tpa_decline_button_tooltip"), ChatColor.GRAY, false));
 
         // Create and send the message
         p.spigot().sendMessage(teleportResponses.create());
@@ -150,4 +157,97 @@ public class TeleportRequestHandler {
         teleportRequests.remove(p);
     }
 
+    public static void startExpiredChecker(Plugin plugin) {
+        Set<Player> expiredTeleportRequests = new HashSet<>();
+        Set<TimedTeleport> completedTeleports = new HashSet<>();
+
+        // Run every second
+        Bukkit.getServer().getScheduler().scheduleSyncRepeatingTask(plugin, () -> {
+
+            // Update active timed teleports
+            if (!TeleportManager.queuedTeleports.isEmpty()) {
+                for (TimedTeleport timedTeleport : TeleportManager.queuedTeleports) {
+                    Player teleporter = Bukkit.getPlayer(timedTeleport.getTeleporter().getUniqueId());
+                    if (teleporter != null) {
+                        if (timedTeleport.getTimeRemaining() > 0) {
+                            if (!timedTeleport.hasMoved(teleporter)) {
+                                if (!timedTeleport.hasLostHealth(teleporter)) {
+                                    teleporter.playSound(teleporter.getLocation(), HuskHomes.settings.getTeleportWarmupSound(), 2, 1);
+                                    MessageManager.sendActionBarMessage(teleporter, "teleporting_action_bar_countdown",
+                                            Integer.toString(timedTeleport.getTimeRemaining()));
+                                    timedTeleport.decrementTimeRemaining();
+                                } else {
+                                    MessageManager.sendActionBarMessage(teleporter, "teleporting_action_bar_cancelled");
+                                    MessageManager.sendMessage(teleporter, "teleporting_cancelled_damage");
+                                    teleporter.playSound(teleporter.getLocation(), HuskHomes.settings.getTeleportCancelledSound(), 1, 1);
+                                    completedTeleports.add(timedTeleport);
+                                }
+                            } else {
+                                MessageManager.sendActionBarMessage(teleporter, "teleporting_action_bar_cancelled");
+                                MessageManager.sendMessage(teleporter, "teleporting_cancelled_movement");
+                                teleporter.playSound(teleporter.getLocation(), HuskHomes.settings.getTeleportCancelledSound(), 1, 1);
+                                completedTeleports.add(timedTeleport);
+                            }
+                        } else {
+                            // Execute the teleport
+                            String targetType = timedTeleport.getTargetType();
+                            switch (targetType) {
+                                case "point":
+                                    TeleportManager.teleportPlayer(teleporter, timedTeleport.getTargetPoint());
+                                    break;
+                                case "player":
+                                    TeleportManager.teleportPlayer(teleporter, timedTeleport.getTargetPlayerName());
+                                    break;
+                                case "random":
+                                    if (HuskHomes.settings.doEconomy()) {
+                                        double rtpCost = HuskHomes.settings.getRtpCost();
+                                        if (rtpCost > 0) {
+                                            if (!VaultIntegration.takeMoney(teleporter, rtpCost)) {
+                                                MessageManager.sendMessage(teleporter, "error_insufficient_funds", VaultIntegration.format(rtpCost));
+                                                break;
+                                            } else {
+                                                MessageManager.sendMessage(teleporter, "rtp_spent_money", VaultIntegration.format(rtpCost));
+                                            }
+                                        }
+                                    }
+                                    TeleportManager.teleportPlayer(teleporter, new RandomPoint(teleporter));
+                                    DataManager.updateRtpCooldown(teleporter);
+                                    break;
+                            }
+                            completedTeleports.add(timedTeleport);
+                        }
+                    } else {
+                        completedTeleports.add(timedTeleport);
+                    }
+                }
+            }
+
+            // Clear completed teleports
+            if (!completedTeleports.isEmpty()) {
+                for (TimedTeleport timedTeleport : completedTeleports) {
+                    TeleportManager.queuedTeleports.remove(timedTeleport);
+                }
+            }
+            completedTeleports.clear();
+
+            // Clear expired teleport requests
+            clearExpiredRequests(expiredTeleportRequests);
+        }, 0L, 20L);
+    }
+
+    // Cancel expired teleport requests
+    private static void clearExpiredRequests(Set<Player> expiredTeleportRequests) {
+        // Check if any requests have expired
+        for (Player p : teleportRequests.keySet()) {
+            if (teleportRequests.get(p).getExpired()) {
+                expiredTeleportRequests.add(p);
+            }
+        }
+
+        // Clear expired requests
+        for (Player p : expiredTeleportRequests) {
+            teleportRequests.remove(p);
+        }
+        expiredTeleportRequests.clear();
+    }
 }
